pagination refactor
